gma tgl: Add port detection
Mostly just detecting the _ISR registers, but during initialization,
this module also blocks TCCOLD so that Modular FIA registers can be
accessed. Instead of keeping track of the state of TCCOLD all of the
time (and enabling and re-enabling), this patch just blocks TCCOLD
once during initializaton and leaves it that way.

Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Change-Id: I674e7c2dcd7738e3d76266ff36675dbd52fa5ae2